# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CurrentUser": {
        "description": "The signed-in user (``GET /auth/me``), also used for the dev login picker.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initials": {
            "description": "One or two letters for the avatar fallback.",
            "readOnly": true,
            "title": "Initials",
            "type": "string"
          },
          "is_platform_admin": {
            "title": "Is Platform Admin",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "display_name",
          "avatar_url",
          "email",
          "is_platform_admin",
          "initials"
        ],
        "title": "CurrentUser",
        "type": "object"
      },
...
      "FieldError": {
        "description": "One request-validation error.",
        "properties": {
          "loc": {
            "description": "Location, e.g. ['body', 'title'].",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem details, returned as ``application/problem+json``.\n\nClients should branch on ``code`` (stable, snake_case), not on ``title`` or\n``detail`` (human-readable, may change). As RFC 9457 allows, ``detail``,\n``instance`` and ``request_id`` may be absent: the one exception to \"every\ndocumented response field is always present\".",
        "examples": [
          {
            "code": "not_found",
            "detail": "Idea not found.",
            "instance": "/api/v1/ideas/0b7c7d1e-7a55-4a4f-9b8b-0d7d3a9d1c11",
            "request_id": "7f9c2b4e1d6a4c0f8e3b5a2d9c1e0f7a",
            "status": 404,
            "title": "Not Found",
            "type": "urn:soundings:problem:not_found"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error code (snake_case).",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation of this occurrence.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path that failed.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlates with server logs (X-Request-ID).",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ProjectRole": {
        "description": "Direct project membership role. Owner/evaluator are per-idea assignments.",
        "enum": [
          "admin",
          "member",
          "viewer"
        ],
        "title": "ProjectRole",
        "type": "string"
      },
...
      "UserPage": {
        "description": "A page of users, ordered by display name.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserSearchResult"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "UserPage",
        "type": "object"
      },
...
      "UserSearchResult": {
        "description": "A user in a people picker (add member, assign owner, invite evaluators).",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initials": {
            "description": "One or two letters for the avatar fallback.",
            "readOnly": true,
            "title": "Initials",
            "type": "string"
          },
          "project_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectRole"
              },
              {
                "type": "null"
              }
            ],
            "description": "The user's role in the project given as ?project=, null if not a member (or no project was given). Only member/admin can be owner or evaluator."
          }
        },
        "required": [
          "id",
          "display_name",
          "avatar_url",
          "email",
          "project_role",
          "initials"
        ],
        "title": "UserSearchResult",
        "type": "object"
      },
      "ValidationProblem": {
        "description": "422 problem with per-field errors (``code`` is ``validation_error``).",
        "examples": [
          {
            "code": "not_found",
            "detail": "Idea not found.",
            "instance": "/api/v1/ideas/0b7c7d1e-7a55-4a4f-9b8b-0d7d3a9d1c11",
            "request_id": "7f9c2b4e1d6a4c0f8e3b5a2d9c1e0f7a",
            "status": 404,
            "title": "Not Found",
            "type": "urn:soundings:problem:not_found"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error code (snake_case).",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation of this occurrence.",
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path that failed.",
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Correlates with server logs (X-Request-ID).",
            "title": "Request Id"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ValidationProblem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "Server-side session cookie set by sign-in (Phase 1: POST /api/v1/auth/dev/login). Unsafe methods must also send the X-CSRF-Token header equal to the soundings_csrf cookie.",
        "in": "cookie",
        "name": "soundings_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Soundings REST API.\n\n* Errors are [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457) problem details\n  (`application/problem+json`); branch on the `code` field.\n* Lists use cursor pagination: pass `next_cursor` back as `?cursor=`.\n* JSON fields are snake_case; timestamps are ISO 8601 with a UTC offset.\n",
    "title": "Soundings API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/dev/users": {
      "get": {
        "description": "Active users, platform admins first, then by name. Development only: 404 unless SOUNDINGS_DEV_LOGIN_ENABLED=true (always refused in production).",
        "operationId": "list_dev_users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CurrentUser"
                  },
                  "title": "Response List Dev Users",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "summary": "List users for the dev login picker",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "description": "Active, non-service-account users whose name or email contains q (case-insensitive), ordered by display name. With project, only members of that project are returned and each result carries project_role.",
        "operationId": "search_users",
        "parameters": [
          {
            "description": "Part of a name or email.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Part of a name or email.",
              "title": "Q"
            }
          },
          {
            "description": "Project slug: members only.",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 48,
                  "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Project slug: members only.",
              "title": "Project"
            }
          },
          {
            "description": "Opaque cursor from next_cursor.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from next_cursor.",
              "title": "Cursor"
            }
          },
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad request (e.g. invalid_cursor)"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not signed in (unauthorized)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found, or not visible to you (not_found)"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "Validation Failed"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Search users",
        "tags": [
          "users"
        ]
      }
    }
  }
}